linux cleanup
git-svn-id: https://svn.code.sf.net/p/oorexx/code-0/interpreter-3.x/trunk@2662 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ kernel/runtime/Numerics.cpp
@@ -135,7 +135,7 @@ RexxObject *Numerics::toObject(int64_t v)
 RexxObject *Numerics::toObject(uint64_t v)
 {
     // in the range for an integer object?
-    if (v <= MAX_WHOLENUMBER)
+    if (v <= (uint64_t)MAX_WHOLENUMBER)
     {
         return new_integer((stringsize_t)v);
     }
